separate payload and vector index bases
