+++ lib/segment/src/index/mod.rs
@@ -1,6 +1,7 @@
 pub mod field_index;
 pub mod hnsw_index;
-mod index_base;
+mod vector_index_base;
+mod payload_index_base;
 mod key_encoding;
 mod payload_config;
 pub mod plain_payload_index;
@@ -10,4 +11,5 @@ mod sample_estimation;
 mod struct_filter_context;
 pub mod struct_payload_index;
 mod visited_pool;
-pub use index_base::*;
+pub use vector_index_base::*;
+pub use payload_index_base::*;
